Split out to clear primitive and complex navigation approaches
