+++ communitytoolkitrepro1/ComplexDetailPage.xaml
@@ -1,0 +1,23 @@
+﻿<?xml version="1.0" encoding="utf-8" ?>
+<ContentPage xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
+             xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
+             xmlns:toolkit="http://schemas.microsoft.com/dotnet/2022/maui/toolkit"
+             xmlns:viewmodel="clr-namespace:communitytoolkitrepro1"
+             x:DataType="viewmodel:ComplexDetailViewModel"
+             x:Class="communitytoolkitrepro1.ComplexDetailPage"
+             Title="DetailPage">
+    <ScrollView>
+        <VerticalStackLayout>
+            <!-- Complex data based query parameter support -->
+            <Label Text="{Binding Path=Character.Name}"
+                   VerticalOptions="Center" 
+                   HorizontalOptions="Center"
+                   FontSize="24" />
+
+            <Button Text="Back"
+                    Command="{Binding GoBackCommand}"
+                    HorizontalOptions="Center"/>
+        </VerticalStackLayout>
+    </ScrollView>
+</ContentPage>
+
