+++ communitytoolkitrepro1/ComplexDetailPage.xaml.cs
@@ -1,0 +1,10 @@
+namespace communitytoolkitrepro1;
+
+public partial class ComplexDetailPage : ContentPage
+{
+	public ComplexDetailPage(ComplexDetailViewModel vm)
+	{
+		InitializeComponent();
+		BindingContext = vm;
+	}
+}
