--- communitytoolkitrepro1/DetailPage.xaml.cs
@@ -1,10 +1,0 @@
-namespace communitytoolkitrepro1;
-
-public partial class DetailPage : ContentPage
-{
-	public DetailPage(DetailViewModel vm)
-	{
-		InitializeComponent();
-		BindingContext = vm;
-	}
-}
