+++ communitytoolkitrepro1/MainViewModel.cs
@@ -3,6 +3,7 @@ using CommunityToolkit.Mvvm.Input;
 using System.Collections.ObjectModel;
 
 namespace communitytoolkitrepro1;
+
 public partial class MainViewModel : ObservableObject
 {
     [ObservableProperty]
@@ -14,19 +15,28 @@ public partial class MainViewModel : ObservableObject
 
         //For repro, just load a few manually
         SubCharacter c = new SubCharacter();
-        c.Name = "Don";
+        c.Name = "PrimitiveData";
         characters.Add(c);
         SubCharacter c2 = new SubCharacter();
-        c2.Name = "Shaun";
+        c2.Name = "ComplexData";
         characters.Add(c2);
     }
 
     [RelayCommand]
     private async Task Tap(Character c)
     {
-        await Shell.Current.GoToAsync(nameof(DetailPage), true, new Dictionary<string, object>()
+        if (c.Name == "PrimitiveData")
         {
-            { "character", c }
-        });
+            // This approach requires the [QueryProperty] usage in the recipient
+            await Shell.Current.GoToAsync($"{nameof(PrimitiveDetailPage)}?name={c.Name}");
+        }
+        else
+        {
+            // This approach required the IQueryAttributable implementation in the recipient
+            await Shell.Current.GoToAsync(nameof(ComplexDetailPage), true, new Dictionary<string, object>()
+            {
+                { "character", c }
+            });
+        }
     }
 }
