+++ communitytoolkitrepro1/PrimitiveDetailPage.xaml.cs
@@ -1,0 +1,10 @@
+namespace communitytoolkitrepro1;
+
+public partial class PrimitiveDetailPage : ContentPage
+{
+	public PrimitiveDetailPage(PrimitiveDetailViewModel vm)
+	{
+		InitializeComponent();
+		BindingContext = vm;
+	}
+}
